% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimators.R
\name{debiased_RRR}
\alias{debiased_RRR}
\title{Title}
\usage{
debiased_RRR(Y, X, r, Sigma_X, sqrt_Gamma = NULL, sqrt_Gamma_inv = NULL)
}
\arguments{
\item{Y}{}

\item{X}{}

\item{r}{}

\item{Sigma_X}{}

\item{sqrt_Gamma}{}

\item{sqrt_Gamma_inv}{}
}
\description{
Title
}
