% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimators.R
\name{debiased_RRR_modified}
\alias{debiased_RRR_modified}
\title{Title}
\usage{
debiased_RRR_modified(
  Y,
  X,
  r,
  Sigma_X,
  ph = NULL,
  sqrt_Gamma = NULL,
  sqrt_Gamma_inv = NULL
)
}
\arguments{
\item{Y}{}

\item{X}{}

\item{r}{}

\item{Sigma_X}{}

\item{ph}{}

\item{sqrt_Gamma}{}

\item{sqrt_Gamma_inv}{}
}
\description{
Title
}
