% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MR_rr_estimators.R
\name{mr_rr_regularized}
\alias{mr_rr_regularized}
\title{Title}
\usage{
mr_rr_regularized(
  Y,
  X,
  r,
  Sigma_X,
  ph = 1e-13,
  sqrt_Gamma = NULL,
  sqrt_Gamma_inv = NULL
)
}
\arguments{
\item{sqrt_Gamma_inv}{}
}
\description{
Title
}
