% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_CI.R
\name{nonpara_bootstrap}
\alias{nonpara_bootstrap}
\title{Title}
\usage{
nonpara_bootstrap(
  me_weight = 0.2,
  regularized = TRUE,
  regularization_rate = 1e-13
)
}
\arguments{
\item{me_weight}{a character value of a positive number, recommend to be chosen between 0.05 and 1. The argument indicates the measurement error weight, which means the scaler we multiply to the measurement error of the coefficients of exposure to SNPs, which is the \eqn{\Sigma_X} in the manuscript. Smaller measurement error weight implies larger IV strength.}

\item{regularized}{can be TRUE or FALSE, indicating whether we use the MR-rr estiomator with the spectral regularization. See the manuscript for more details.}

\item{regularization_rate}{a small positive numerical value, indicating the regularization rate. It is only used when regularized = TRUE. The larger regularization_rate make the estimator having smaller variance, but the rate can not be too large to ensure consistency holds, see more details in the manuscript.}
}
\value{
The average coverage rate of the entry-wise 95\% confidence interval produced by non-parametric bootstrap method.
}
\description{
Title
}
\examples{
nonpara_bootstrap(me_weight = 0.2, regularized = TRUE, regularization_rate = 1e-13)
}
