% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MR_rr_simulation.R
\name{plot_boxplot}
\alias{plot_boxplot}
\title{The function to create the boxplot for the bias of the naive MR-rr estimator and the MR-rr estimator (with spectral regularization) comparing to true C for each entry.}
\usage{
plot_boxplot(
  result_AB_list,
  result_AB_d_list,
  weight_to_plot,
  individual_plot = FALSE,
  rank_by = "bias"
)
}
\arguments{
\item{result_AB_list}{a numerical list of the simulated bias of the naive MR-rr estimator comparing to true C for each entry. It is the first output of the function "run_simulation".}

\item{result_AB_d_list}{a numerical list of the bias of the simulated MR-rr estimator comparing to true C for each entry. It is the second output of the function "run_simulation".}

\item{weight_to_plot}{a character value that can be chosen from "1", "0.5", "0.2", "0.1", "0.05", indicating the plot we want to make under what measurement error weight. The weight means the scaler we multiply to the measurement error of the coefficients of exposure to SNPs, which is the \eqn{\Sigma_X} in the manuscript. Smaller measurement error weight implies larger IV strength.}

\item{individual_plot}{a logical value TRUE or FALSE. TRUE indicates to plot the boxplot for entries of C separately. FALSE indicates to crate combined plot including both estimators for some entries of C. The Default is FALSE.}

\item{rank_by}{a character value that can be chosen from "bias" or "sd". It indicates the criteria we use to rank the entries to plot. "bias" means we rank the entries by the average bias of each estimator, and "sd" means we rank the entries by the average standard deviation. The Default is "bias".}
}
\value{
a list of the boxplot of the entries with largest bias or standard deviation of the naive MR-rr estimator and the MR-rr estimator. Return either a combined plot with 5 entries or 3 separately plots depend on the input "individual_plot".
}
\description{
The function to create the boxplot for the bias of the naive MR-rr estimator and the MR-rr estimator (with spectral regularization) comparing to true C for each entry.
}
