% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MR_rr_simulation.R
\name{plot_heatmap}
\alias{plot_heatmap}
\title{The function to generate the heatmap of the average absolute bias and the standard deviation of the naive MR-rr estimator and the MR-rr estimator (with spectral regularization).}
\usage{
plot_heatmap(result_AB_list, result_AB_d_list, weight_to_plot)
}
\arguments{
\item{result_AB_list}{a numerical list of the simulated bias of the naive MR-rr estimator comparing to true C for each entry. It is the first output of the function "run_simulation".}

\item{result_AB_d_list}{a numerical list of the bias of the simulated MR-rr estimator comparing to true C for each entry. It is the second output of the function "run_simulation".}

\item{weight_to_plot}{a character value that can be chosen from "1", "0.5", "0.2", "0.1", "0.05", indicating the plot we want to make under what measurement error weight. The weight means the scaler we multiply to the measurement error of the coefficients of exposure to SNPs, which is the \eqn{\Sigma_X} in the manuscript. Smaller measurement error weight implies larger IV strength.}
}
\value{
four heatmaps containing the entry-wise average absolute bias and the standard deviation of the naive MR-rr estimator and the MR-rr estimator. Also return a list of numeric values including the IV strength under this measurement error weight, the mean absolute value of the entries of C, the average absolute bias of the naive MR-rr estimator and the MR-rr estimator, the standard deviation of the naive MR-rr estimator and the MR-rr estimator.
}
\description{
The function to generate the heatmap of the average absolute bias and the standard deviation of the naive MR-rr estimator and the MR-rr estimator (with spectral regularization).
}
