% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimators.R
\name{RRR}
\alias{RRR}
\title{Title}
\usage{
RRR(Y, X, r, sqrt_Gamma = NULL, sqrt_Gamma_inv = NULL)
}
\arguments{
\item{Y}{}

\item{X}{}

\item{r}{}

\item{sqrt_Gamma}{}

\item{sqrt_Gamma_inv}{}
}
\description{
Title
}
