% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MR_rr_simulation.R
\name{run_simulation}
\alias{run_simulation}
\title{Title}
\usage{
run_simulation(regularized = TRUE, regularization_rate = 1e-13)
}
\arguments{
\item{regularized}{can be TRUE or FALSE, indicating whether we use the MR-rr estiomator with the spectral regularization. See the manuscript for more details.}

\item{regularization_rate}{a small positive numerical value, indicating the regularization rate. It is only used when regularized = TRUE. The larger regularization_rate make the estimator having smaller variance, but the rate can not be too large to ensure consistency holds, see more details in the manuscript.}
}
\value{
a nested list of two list element for the simulation results, where the first list contains the bias of the naive MR-rr estimator comparing to true C for each entry; the second list contains the bias of the MR-rr estimator comparing to true C for each entry. For either bias list, the list have 5 list elements, corresponding to the measurement error weights (see more description in the "plot_boxplot" function) 1, 0.5, 0.2, 0.1, 0.05. Each element is 240 by 100, which is the number of entries and the iteration in the simulation.
}
\description{
Title
}
\examples{
result_AB_list = run_simulation(regularized = TRUE, regularization_rate = 1e-13)[[1]]
result_AB_d_list = run_simulation(regularized = TRUE, regularization_rate = 1e-13)[[2]]
str(result_AB_list)
summary(unlist(result_AB_list))
summary(unlist(result_AB_d_list))
}
